requirements.txt
confluent-kafka==2.4.0
groq
python-dotenv
redis